+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Fix: Log actual payloads by default (was logging "null" for the payload).
+
 ## 2.0.13 - 2025-05-26
 
 - Fix: Outbox producer would crash if no Deimos producer was defined for the topic.